used workaround

+++ Peregrine/run_tests_ci.sh
@@ -5,15 +5,16 @@ for file in *
 do
 echo $file ...
 ../peregrine compile $file -emit-c
-gcc temp.c -o output.o | grep 'error' -q
-if [$? -ne 0]
-then
-echo "::error GCC FAILED"
-exit 1
-fi
+gcc temp.c -o output.o 
+if test -f "output.o";then
 chmod +x ./output.o
 ./output.o
 rm ./output.o
+else
+echo "::error GCC FAILED"
+exit 1
+fi
+
 
     
 done
